requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
dataclasses==0.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-rest-swagger==2.2.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-summernote==0.8.20.0
django-templated-mail==1.1.1
djangorestframework==3.13.0
djangorestframework-jsonapi==4.3.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
djangorestframework-stubs==1.4.0
djoser==2.1.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.6
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
types-pytz==2021.3.2
types-PyYAML==6.0.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.3.0
